services/embedder/requirements.txt
fastapi
uvicorn
pydantic

google-cloud-bigquery
google-cloud-aiplatform
